polimento e funcionamento do código
