+++ Tests/ASPS/CriarEmendaDeValorAoPLOA2024Test.java
@@ -1,36 +1,26 @@
 // Generated by Selenium IDE
 import org.junit.Test;
 import org.junit.Before;
+import org.hamcrest.MatcherAssert;
 import org.junit.After;
-import static org.junit.Assert.*;
 import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.core.IsNot.not;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.remote.RemoteWebDriver;
-import org.openqa.selenium.remote.DesiredCapabilities;
-import org.openqa.selenium.Dimension;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.Alert;
-import org.openqa.selenium.Keys;
 import java.util.*;
-import java.net.MalformedURLException;
-import java.net.URL;
+import java.time.Duration;
 public class CriarEmendaDeValorAoPLOA2024Test {
   private WebDriver driver;
-  private Map<String, Object> vars;
   JavascriptExecutor js;
   @Before
   public void setUp() {
     driver = new ChromeDriver();
     js = (JavascriptExecutor) driver;
-    vars = new HashMap<String, Object>();
+    new HashMap<String, Object>();
   }
   @After
   public void tearDown() {
@@ -40,8 +30,24 @@ public class CriarEmendaDeValorAoPLOA2024Test {
   public void criarEmendaDeValorAoPLOA2024() {
     // Abrir site
     driver.get("http://10.2.1.110:8080/legislativo-proposta-emenda-loa/2024/gabinete");
+    driver.findElement(By.id("kc-form-login")).click();
+    driver.findElement(By.id("username")).click();
+    driver.findElement(By.id("username")).sendKeys("usuario");
+    driver.findElement(By.id("password")).sendKeys("usuario");
+    driver.findElement(By.id("kc-login")).click();
+    {
+      WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(1020, 1));
+      wait.until(ExpectedConditions.presenceOfElementLocated(By.cssSelector("#jh-create-entity > span")));
+    }
+    {
+      List<WebElement> elements = driver.findElements(By.cssSelector("#jh-create-entity > span"));
+      assert(elements.size() > 0);
+    }
     driver.findElement(By.cssSelector("#jh-create-entity > span")).click();
     // Verificar se está na página certa se guiando pelo texto
-    assertThat(driver.findElement(By.id("jhi-legislativo-proposta-emenda-loa-heading")).getText(), is("Criar/Editar Proposta de Emenda ao PLOA 2024"));
+
+    WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(1020, 1));
+      wait.until(ExpectedConditions.presenceOfElementLocated(By.id("jhi-legislativo-proposta-emenda-loa-heading")));
+    MatcherAssert.assertThat(driver.findElement(By.id("jhi-legislativo-proposta-emenda-loa-heading")).getText(), is("Criar/Editar Remanejamento ao PLOA 2024"));
   }
 }
