+++ Tests/ASPS/SelecionarAçãoTest.java
@@ -22,7 +22,7 @@ import org.openqa.selenium.Keys;
 import java.util.*;
 import java.net.MalformedURLException;
 import java.net.URL;
-public class SelecionarAoTest {
+public class SelecionarAçãoTest {
   private WebDriver driver;
   private Map<String, Object> vars;
   JavascriptExecutor js;
